Extract reusable code to helper module
